# starlark: Reduce allocations for `find`/`indexof` without `end`