By using the overload of `String#indexOf` that accepts a starting position, the common case of Starlark string forward searches that extend to the end of the string no longer requires allocations.

Before:
```
benchmark                        ops     cpu/op    wall/op   steps/op   alloc/op
bench_find                  33554431      223ns      221ns          6       143B
bench_find_with_start       33554431      238ns      236ns          6       631B
```

After:
```
benchmark                        ops     cpu/op    wall/op   steps/op   alloc/op
bench_find                  33554431      221ns      219ns          6       143B
bench_find_with_start       33554431      210ns      210ns          6       143B
```